+++ src/push/type/item/scalar.clj
@@ -261,21 +261,24 @@
 (def scalar-power
   (core/build-instruction
     scalar-power
-    "`:scalar-power` pops the top two `:scalar` values (call them `exponent` and `base` respectively). It calculates `(numeric-tower/expt base exponent)`. If the absolute value of the product of the exponent and `(Math/log base)` is more than 2^16, no result is returned; instead, an `:error` is pushed."
+    "`:scalar-power` pops the top two `:scalar` values (call them `exponent` and `base` respectively). It calculates `(numeric-tower/expt base exponent)`. If the absolute value of the product of the exponent and `(Math/log base)` is more than 2^16, no result is returned; instead, an `:error` is pushed. Unlike most `:scalar` instructions, if the value is positive or negative `Infinity`, an error is also produced."
     :tags #{:arithmetic :base :dangerous}
 
     (d/consume-top-of :scalar :as :exp)
     (d/consume-top-of :scalar :as :base)
     (d/calculate [:base :exp]
-      #(if (zero? %1)
-        false
-        (> (nt/abs (*' %2 (Math/log (nt/abs %1)))) 
-           65535)) :as :oversized?)
+      #(cond (zero? %1) false
+             (math/infinite? %1) false
+             (math/infinite? %2) false
+             (> (nt/abs (*' %2 (Math/log (nt/abs %1)))) 65535) true
+             :else false) :as :oversized?)
     (d/calculate [:base :exp :oversized?]
       #(if %3 nil (nt/expt %1 %2)) :as :prelim)
     (d/calculate [:prelim]
       #(or (nil? %1)
-           (Double/isNaN %1)) :as :bad-result)
+           (math/infinite? %1)
+           (Double/isNaN %1)
+           ) :as :bad-result)
     (d/calculate [:bad-result :prelim]
       #(if %1 nil %2) :as :result)
     (d/calculate [:bad-result]