+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - `ref-dump-tagspace` instruction
 - renamed `collectible` aspect to `set-able` throughout
 - imposed a restriction on `:scalar-power`: produces an `:error` when `(abs (*' base (Math/log expt)))` is larger than 65535 (that magic number could change after stress testing), to avoid time-out errors in random programs generating huuuuuuuuuge results blithely
+- minor change to `:scalar-power`, which will produce an `:error` if the result is `∞` or `-∞` (unlike most other instructions)
 
 ## 0.1.21